- code clean up: unnecessary signed and unsigned  - cast changes

+++ FaxModulator.cpp
@@ -22,10 +22,9 @@ FaxModulator::FaxModulator(QObject* parent)
 	: QObject(parent), sampleRate(0), carrier(0), dev(0)
 {
 	// put complete period of sine in lookup table
-	sine=new signed short[size_sine];
-	for(unsigned int i=0; i<size_sine; i++) {
-		sine[i]=(signed short)
-			(32767*sin(2.*M_PI*(double)i/(double)size_sine));
+	sine=new short[size_sine];
+	for(int i=0; i<size_sine; i++) {
+		sine[i]=static_cast<short>(32767*sin(2.0*M_PI*i/size_sine));
 	}
 }
 
@@ -34,17 +33,17 @@ FaxModulator::~FaxModulator(void)
 	delete sine;
 }
 
-void FaxModulator::modulate(double* buffer, unsigned int number)
+void FaxModulator::modulate(double* buffer, int number)
 {
-	signed short sample[number];
-	for(unsigned int i=0; i<number; i++) {
+	short sample[number];
+	for(int i=0; i<number; i++) {
 		if(fm) {
 			sample[i]=sine[phase];
-			unsigned int f=static_cast<unsigned int>
+			int f=static_cast<int>
 				(carrier+2.0*(buffer[i]-0.5)*dev);
 			phase+=size_sine*f/sampleRate;
 		} else {
-			sample[i]=static_cast<signed short>
+			sample[i]=static_cast<short>
 				(0.04+0.96*buffer[i]*sine[phase]);
 			phase+=size_sine*carrier/sampleRate;
 		}
